+++ CHANGELOG.md
@@ -26,7 +26,9 @@
 
 
 ## 1.7.1-SNAPSHOT
-go ブロックは呼んだ順に実行されるとは限らない。
+- go ブロックは呼んだ順に実行されるとは限らない。
+- reset! を swap! に戻す。
+- defonce を１つに。除く setInterval.
 
 ## 1.7.0 - 2022-05-12
 code polish up. not improved.